Updated vulnscan.py
Now it incorporates tqdm to show progress of saving file data, also setup files to be used now for logicytics
