+++ .idea/Logicytics.iml
@@ -8,8 +8,8 @@
       <sourceFolder url="file://$MODULE_DIR$/CODE/libs" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/CODE/_internal" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/TESTS" isTestSource="true" />
-      <sourceFolder url="file://$MODULE_DIR$/CODE/VulnScan/v2" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/CODE/VulnScan/v3" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/CODE/VulnScan/v2-deprecated" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/ACCESS" />
       <excludeFolder url="file://$MODULE_DIR$/CODE/VulnScan/cves" />
       <excludeFolder url="file://$MODULE_DIR$/CODE/VulnScan/albert" />
