--- CODE/VulnScan/tools/test_gpu_acceleration.py
@@ -1,11 +1,0 @@
-import torch
-
-
-def check_gpu():
-    if torch.cuda.is_available():
-        print(f"CUDA is available. Using GPU: {torch.cuda.get_device_name(0)}")
-    else:
-        print("CUDA is not available. Using CPU.")
-
-
-check_gpu()
