+++ CODE/config.ini
@@ -74,8 +74,9 @@ partial_sensitive_chance = 0.2
 # The following settings are for the Vectorizer module for vectorizing data
 # Usually it automatically vectorizes data, but this is for manual vectorization
 
-# We advise to use the automatic vectorization,
+# We advise to use this vectorization, although not knowing the vectorizer is not advised
 # as this may lead to ValueErrors due to different inputs
+# Use the vectorizer supplied for any v3 model on SenseMini
 
 # The path to the data to vectorize, either a file or a directory
 data_path = C:\Users\Hp\Desktop\Model Tests\Model Data\GeneratedData
